requirements.txt
torch>=2.0.0
transformers>=4.30.0
diffusers>=0.21.0
accelerate>=0.20.0
pillow>=9.0.0
numpy>=1.24.0
opencv-python>=4.8.0
svgwrite>=1.4.0
cairosvg>=2.7.0
matplotlib>=3.7.0
scikit-image>=0.21.0
gradio>=4.0.0
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.31.0
